+++ zerorvc/synthesizer/commons.py
@@ -66,7 +66,7 @@ def rand_slice_segments(x: torch.Tensor, x_lengths=None, segment_size=4):
     if x_lengths is None:
         x_lengths = t
     ids_str_max = x_lengths - segment_size + 1
-    ids_str = (torch.rand([b]).to(device=x.device) * ids_str_max).to(dtype=torch.long)
+    ids_str = (torch.rand([b]).to(device=x.device) * ids_str_max).to(dtype=torch.int32)
     ret = slice_segments(x, ids_str, segment_size)
     return ret, ids_str
